Review: Extract JavaClassDependencies, since JavaClass has too many mixed concerns already
Issue: #136
